requirements.txt
black==23.3.0
flake8==6.0.0
Pypubsub==4.0.3
pyserial==3.5
pyserial-asyncio==0.6
pytest==7.3.1
pytest-asyncio==0.21.0
pytest-cov==4.0.0
pytest-mock==3.10.0
python-dateutil==2.8.2
python-statemachine==2.0.0
pytz==2023.3
PyYAML==6.0
requests==2.30.0
retry==0.9.2
six==1.16.0